[nimble] Making sun compiler happy
git-svn-id: https://gridkit.googlecode.com/svn/grid-lab/trunk/nimble@2289 4e96dbfe-9890-11de-ac5d-73b9d092dc0a

+++ probe-jmx/src/main/java/org/gridkit/lab/util/jmx/mxstruct/coherence/ConnectionManagerMXStruct.java
@@ -11,7 +11,7 @@ public class ConnectionManagerMXStruct extends MXStruct {
 
 	@AttrName("ConnectionCount")
 	public int getConnectionCount() {
-		return super.getMXAttr();
+		return (Integer)super.getMXAttr();
 	}
 
 	@AttrName("HostIP")
